server/requirements.txt
Flask
flask-cors
numpy
pandas
scikit-learn
python-dotenv
pickle-mixin
gunicorn
pickle